requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
python-dotenv
sqlalchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
langchain>=0.2.10
langchain-openai
langchain-community
chromadb
tiktoken
loguru
structlog
httpx
pandas
openpyxl
langchain-text-splitters
httpx
pytest
pytest-asyncio
langchain-text-splitters
pypdf
